Moves operations to separate files
